pyproject.toml
[project]
name = "employee-server"
version = "0.4.0"
description = "Employee registry API server"
authors = [
  { name = "Toplica Tanasković", email = "toplicius@gmail.com" },
]
dependencies = [
  "Flask",
  "Flask-SQLAlchemy",
  "flask-restx",
]
requires-python = ">3.8"

[project.optional-dependencies]
test = [
  "pytest",
  "coverage",
  "toml",
]

[build-system]
requires = ["flit-core"]
build-backend = "flit_core.buildapi"

[tool.coverage.run]
branch = true
source = [ 
  "employee_server", 
]

[tool.black]
line-length = 79
